Fix type error: cast auth()->id() to int in duplicates page

+++ resources/views/livewire/admin/duplicates/index.blade.php
@@ -19,7 +19,7 @@ $search = function () {
     $this->duplicates = [];
 
     $service = app(DuplicateDetectionService::class);
-    $userId = auth()->id();
+    $userId = (int) auth()->id();
 
     $results = match ($this->modelType) {
         'Event' => $service->findDuplicateEvents($userId, $this->threshold, $this->limit),
